+++ src/main/java/backend/housekeeper/houseelegantservice/domain/model/command/CreateHouseCommand.java
@@ -1,5 +1,5 @@
 package backend.housekeeper.houseelegantservice.domain.model.command;
 
 public record CreateHouseCommand(String country, String city, String streetAddress,
-                                 double price, double rating, String photoUrl, int capacity) {
+                                 Float price, Float rating, String photoUrl, Integer capacity) {
 }
